mobileclip_mobile/requirements.txt
# MobileCLIP Mobile Deployment Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0
pyyaml>=6.0
tqdm>=4.65.0
psutil>=5.9.0

# MobileCLIP (install from GitHub)
# Run: pip install git+https://github.com/apple/ml-mobileclip.git

# Mobile conversion frameworks
# iOS conversion (macOS only)
coremltools>=7.0; sys_platform == "darwin"

# Android conversion
tensorflow>=2.13.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Optional: ONNX-TF for TensorFlow Lite conversion
# onnx-tf>=1.10.0

# Model downloading and management
huggingface-hub>=0.16.0
transformers>=4.30.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Documentation (optional)
# sphinx>=5.0.0
# sphinx-rtd-theme>=1.2.0
